+++ domain/src/Aggregates/NonFungibleAsset/Repositories/NonFungibleAssetRepository.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Domain\Aggregates\NonFungibleAsset\Repositories;
 
 use Domain\Aggregates\NonFungibleAsset\NonFungibleAsset;
-use Domain\Aggregates\NonFungibleAsset\NonFungibleAssetId;
+use Domain\Aggregates\NonFungibleAsset\ValueObjects\NonFungibleAssetId;
 
 interface NonFungibleAssetRepository
 {
