+++ domain/src/Aggregates/TaxYear/Repositories/TaxYearMessageRepository.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Domain\Aggregates\TaxYear\Repositories;
 
-use Domain\Aggregates\TaxYear\TaxYearId;
+use Domain\Aggregates\TaxYear\ValueObjects\TaxYearId;
 use Generator;
 
 interface TaxYearMessageRepository
