+++ domain/src/Aggregates/TaxYear/Repositories/TaxYearRepository.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Domain\Aggregates\TaxYear\Repositories;
 
 use Domain\Aggregates\TaxYear\TaxYear;
-use Domain\Aggregates\TaxYear\TaxYearId;
+use Domain\Aggregates\TaxYear\ValueObjects\TaxYearId;
 
 interface TaxYearRepository
 {
