+++ domain/src/Services/TransactionDispatcher/Handlers/IncomeHandler.php
@@ -6,7 +6,7 @@ namespace Domain\Services\TransactionDispatcher\Handlers;
 
 use Domain\Aggregates\TaxYear\Actions\UpdateIncome;
 use Domain\Aggregates\TaxYear\Repositories\TaxYearRepository;
-use Domain\Aggregates\TaxYear\TaxYearId;
+use Domain\Aggregates\TaxYear\ValueObjects\TaxYearId;
 use Domain\Services\TransactionDispatcher\Handlers\Exceptions\IncomeHandlerException;
 use Domain\ValueObjects\Transactions\Acquisition;
