+++ domain/tests/Aggregates/NonFungibleAsset/Reactors/NonFungibleAssetReactorTestCase.php
@@ -2,7 +2,7 @@
 
 namespace Domain\Tests\Aggregates\NonFungibleAsset\Reactors;
 
-use Domain\Aggregates\NonFungibleAsset\NonFungibleAssetId;
+use Domain\Aggregates\NonFungibleAsset\ValueObjects\NonFungibleAssetId;
 use Domain\Aggregates\NonFungibleAsset\Reactors\NonFungibleAssetReactor;
 use Domain\Aggregates\TaxYear\Repositories\TaxYearRepository;
 use EventSauce\EventSourcing\MessageConsumer;
