+++ domain/tests/Aggregates/SharePoolingAsset/Reactors/SharePoolingAssetReactorTestCase.php
@@ -3,7 +3,7 @@
 namespace Domain\Tests\Aggregates\SharePoolingAsset\Reactors;
 
 use Domain\Aggregates\SharePoolingAsset\Reactors\SharePoolingAssetReactor;
-use Domain\Aggregates\SharePoolingAsset\SharePoolingAssetId;
+use Domain\Aggregates\SharePoolingAsset\ValueObjects\SharePoolingAssetId;
 use Domain\Aggregates\TaxYear\Repositories\TaxYearRepository;
 use EventSauce\EventSourcing\MessageConsumer;
 use EventSauce\EventSourcing\TestUtilities\MessageConsumerTestCase;
