+++ domain/tests/Aggregates/TaxYear/Factories/Projections/TaxYearSummaryFactory.php
@@ -4,7 +4,7 @@ namespace Domain\Tests\Aggregates\TaxYear\Factories\Projections;
 
 use Brick\DateTime\LocalDate;
 use Domain\Aggregates\TaxYear\Projections\TaxYearSummary;
-use Domain\Aggregates\TaxYear\TaxYearId;
+use Domain\Aggregates\TaxYear\ValueObjects\TaxYearId;
 use Domain\Aggregates\TaxYear\ValueObjects\CapitalGain;
 use Domain\Enums\FiatCurrency;
 use Domain\ValueObjects\FiatAmount;
